Refactor to move keycloak into a package.
Move keycloak code to a package. And use faktory.

Issue #HAV-133
