+++ havenapi/actions/registration_funnel.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"strings"
 
+	faktory "github.com/contribsys/faktory/client"
 	"github.com/deis/helm/log"
 	"github.com/gobuffalo/buffalo"
 	"github.com/gobuffalo/pop"
@@ -32,6 +33,14 @@ func RegistrationHandler(c buffalo.Context) error {
 				"Error inserting registration to database: %s for remote address %s",
 				err.Error(),
 				remoteAddress))
+	}
+
+	// Add job to the queue
+	client, err := faktory.Open()
+	job := faktory.NewJob("CreateUser", request.FormValue("email"))
+	err = client.Push(job)
+	if err != nil {
+		return c.Error(500, err)
 	}
 
 	log.Info("processed a registration")
