+++ havenapi/keycloak/keycloak_test.go
@@ -1,4 +1,4 @@
-package actions
+package keycloak
 
 func (as *ActionSuite) Test_KeycloakGetToken() {
 	err := KeycloakGetToken()
